requirements.txt
streamlit
python-docx
docx2python
openpyxl